+++ python/marvin/utils/general/general.py
@@ -21,6 +21,7 @@ import re
 
 from collections import OrderedDict
 from builtins import range
+from functools import wraps
 
 import numpy as np
 
@@ -36,7 +37,8 @@ import marvin
 
 from marvin import log
 from marvin.core.exceptions import MarvinError, MarvinUserWarning
-from marvin.utils.datamodel.dap.plotting import get_default_plot_params
+from brain.core.exceptions import BrainError
+from flask_jwt_extended import get_jwt_identity
 
 try:
     from sdss_access import RsyncAccess, AccessError
@@ -61,12 +63,28 @@ except ImportError as e:
 __all__ = ('convertCoords', 'parseIdentifier', 'mangaid2plateifu', 'findClosestVector',
            'getWCSFromPng', 'convertImgCoords', 'getSpaxelXY',
            'downloadList', 'getSpaxel', 'get_drpall_row', 'getDefaultMapPath',
-           'getDapRedux', 'get_nsa_data', '_check_file_parameters', 'get_plot_params',
+           'getDapRedux', 'get_nsa_data', '_check_file_parameters',
            'invalidArgs', 'missingArgs', 'getRequiredArgs', 'getKeywordArgs',
            'isCallableWithArgs', 'map_bins_to_column', '_sort_dir',
-           'get_dapall_file', 'temp_setattr', 'map_dapall', 'turn_off_ion', 'memory_usage')
+           'get_dapall_file', 'temp_setattr', 'map_dapall', 'turn_off_ion', 'memory_usage',
+           'validate_jwt', 'target_status', 'target_is_observed')
 
 drpTable = {}
+
+
+def validate_jwt(f):
+    ''' Decorator to validate a JWT and User '''
+
+    @wraps(f)
+    def wrapper(*args, **kwargs):
+        current_user = get_jwt_identity()
+
+        if not current_user:
+            raise MarvinError('Invalid user from API token!')
+        else:
+            marvin.config.access = 'collab'
+        return f(*args, **kwargs)
+    return wrapper
 
 
 def getSpaxel(cube=True, maps=True, modelcube=True,
@@ -476,7 +494,7 @@ def findClosestVector(point, arr_shape=None, pixel_shape=None, xyorig=None):
     return minind
 
 
-def getWCSFromPng(image):
+def getWCSFromPng(filename=None, image=None):
     """Extract any WCS info from the metadata of a PNG image.
 
     Extracts the WCS metadata info from the PNG optical
@@ -484,7 +502,9 @@ def getWCSFromPng(image):
     Converts it to an Astropy WCS object.
 
     Parameters:
-        image (str):
+        image (object):
+            An existing PIL image object
+        filename (str):
             The full path to the image
 
     Returns:
@@ -492,11 +512,18 @@ def getWCSFromPng(image):
             an Astropy WCS object
     """
 
+    assert any([image, filename]), 'Must provide either a PIL image object, or the full image filepath'
+
     pngwcs = None
-    try:
-        image = PIL.Image.open(image)
-    except Exception as e:
-        raise MarvinError('Cannot open image {0}: {1}'.format(image, e))
+
+    if filename and not image:
+        try:
+            image = PIL.Image.open(filename)
+        except Exception as e:
+            raise MarvinError('Cannot open image {0}: {1}'.format(filename, e))
+        else:
+            # Close the image
+            image.close()
 
     # get metadata
     meta = image.info if image else None
@@ -516,9 +543,6 @@ def getWCSFromPng(image):
     # Construct Astropy WCS
     if mywcs:
         pngwcs = wcs.WCS(mywcs)
-
-    # Close the image
-    image.close()
 
     return pngwcs
 
@@ -651,7 +675,9 @@ def getDapRedux(release=None):
     if not Path:
         raise MarvinError('sdss_access is not installed')
     else:
-        sdss_path = Path()
+        is_public = 'DR' in release
+        path_release = release.lower() if is_public else None
+        sdss_path = Path(public=is_public, release=path_release)
 
     release = release or marvin.config.release
     drpver, dapver = marvin.config.lookUpVersions(release=release)
@@ -686,11 +712,6 @@ def getDefaultMapPath(**kwargs):
             The sas url to download the default maps file
     """
 
-    if not Path:
-        raise MarvinError('sdss_access is not installed')
-    else:
-        sdss_path = Path()
-
     # Get kwargs
     release = kwargs.get('release', marvin.config.release)
     drpver, dapver = marvin.config.lookUpVersions(release=release)
@@ -698,6 +719,14 @@ def getDefaultMapPath(**kwargs):
     ifu = kwargs.get('ifu', None)
     daptype = kwargs.get('daptype', 'SPX-GAU-MILESHC')
     bintype = kwargs.get('bintype', 'MAPS')
+
+    # get sdss_access Path
+    if not Path:
+        raise MarvinError('sdss_access is not installed')
+    else:
+        is_public = 'DR' in release
+        path_release = release.lower() if is_public else None
+        sdss_path = Path(public=is_public, release=path_release)
 
     # get the sdss_path name by MPL
     # TODO: this is likely to break in future MPL/DRs. Just a heads up.
@@ -806,8 +835,12 @@ def downloadList(inputlist, dltype='cube', **kwargs):
     elif dltype == 'image':
         name = 'mangaimage'
 
+    # check for public release
+    is_public = 'DR' in release
+    rsync_release = release.lower() if is_public else None
+
     # create rsync
-    rsync_access = RsyncAccess(label='marvin_download', verbose=verbose)
+    rsync_access = RsyncAccess(label='marvin_download', verbose=verbose, public=is_public, release=rsync_release)
     rsync_access.remote()
 
     # Add objects
@@ -1019,20 +1052,6 @@ def _check_file_parameters(obj1, obj2):
                       .format(param, obj1.__repr__, obj2.__repr__, getattr(obj1, param),
                               getattr(obj2, param)))
         assert getattr(obj1, param) == getattr(obj2, param), assert_msg
-
-
-def get_plot_params(dapver, prop):
-    """Return default plotting parameters for a property."""
-    params = get_default_plot_params(dapver)
-
-    if 'vel' in prop:
-        key = 'vel'
-    elif 'sigma' in prop:
-        key = 'sigma'
-    else:
-        key = 'default'
-
-    return params[key]
 
 
 def add_doc(value):
@@ -1284,7 +1303,11 @@ def get_dapall_file(drpver, dapver):
 
     assert Path is not None, 'sdss_access.path.Path is not available.'
 
-    dapall_path = Path().full('dapall', dapver=dapver, drpver=drpver)
+    release = marvin.config.lookUpRelease(drpver)
+    is_public = 'DR' in release
+    path_release = release.lower() if is_public else None
+    path = Path(public=is_public, release=path_release)
+    dapall_path = path.full('dapall', dapver=dapver, drpver=drpver)
 
     return dapall_path
 
@@ -1493,3 +1516,67 @@ def memory_usage(where):
     print("Memory summary: {0}".format(where))
     pympler.summary.print_(mem_summary, limit=2)
     print("VM: {0:.2f}Mb".format(get_virtual_memory_usage_kb() / 1024.0))
+
+
+def target_status(mangaid, mode='auto', source='nsa', drpall=None, drpver=None):
+    ''' Check the status of a MaNGA target
+
+    Given a mangaid, checks the status of a target.  Will check if
+    target exists in the NSA catalog (i.e. is a proper target) and checks if
+    target has a corresponding plate-IFU designation (i.e. has been observed).
+
+    Returns a string status indicating if a target has been observed, has not
+    yet been observed, or is not a valid MaNGA target.
+
+    Parameters:
+        mangaid (str):
+            The mangaid of the target to check for observed status
+        mode ({'auto', 'drpall', 'db', 'remote', 'local'}):
+            See mode in :func:`mangaid2plateifu` and :func:`get_nsa_data`.
+        source ({'nsa', 'drpall'}):
+            The NSA catalog data source. See source in :func:`get_nsa_data`.
+        drpall (str or None):
+            The drpall file to use.  See drpall in :func:`mangaid2plateifu` and :func:`get_nsa_data`.
+        drpver (str or None):
+            The DRP version to use.  See drpver in :func:`mangaid2plateifu` and :func:`get_nsa_data`.
+
+    Returns:
+        A status of "observed", "not yet observed", or "not valid target"
+
+    '''
+
+    # check for plateifu - target has been observed
+    try:
+        plateifu = mangaid2plateifu(mangaid, mode=mode, drpver=drpver, drpall=drpall)
+    except (MarvinError, BrainError) as e:
+        plateifu = None
+
+    # check if target in NSA catalog - proper manga target
+    try:
+        nsa = get_nsa_data(mangaid, source=source, mode=mode, drpver=drpver, drpall=drpall)
+    except (MarvinError, BrainError) as e:
+        nsa = None
+
+    # return observed boolean
+    if plateifu and nsa:
+        status = 'observed'
+    elif not plateifu and nsa:
+        status = 'not yet observed'
+    elif not plateifu and not nsa:
+        status = 'not valid target'
+
+    return status
+
+
+def target_is_observed(mangaid, mode='auto', source='nsa', drpall=None, drpver=None):
+    ''' Check if a MaNGA target has been observed or not
+
+    See :func:`target_status` for full documentation.
+
+    Returns:
+        True if the target has been observed.
+
+    '''
+    # check the target status
+    status = target_status(mangaid, source=source, mode=mode, drpver=drpver, drpall=drpall)
+    return status == 'observed'