+++ python/marvin/web/controllers/index.py
@@ -4,17 +4,12 @@ from flask_classful import route
 from marvin import config, marvindb
 from brain.api.base import processRequest
 from marvin.utils.general.general import parseIdentifier
-from marvin.web.web_utils import parseSession
 from marvin.api.base import arg_validate as av
 from marvin.web.controllers import BaseWebView
+from marvin.web.web_utils import setGlobalSession, set_session_versions
 
-from hashlib import md5
-# try:
-#     print('importing main inspection')
-#     from inspection.marvin import Inspection
-# except ImportError as e:
-#     print('importing local inspection')
-#     from brain.core.inspection import Inspection
+from brain.utils.general import validate_user, get_db_user
+from flask_login import current_user, login_user, logout_user
 
 index = Blueprint("index_page", __name__)
 
@@ -32,7 +27,6 @@ class Marvin(BaseWebView):
 
     def index(self):
         current_app.logger.info('Welcome to Marvin Web!')
-
         return render_template("index.html", **self.main)
 
     def quote(self):
@@ -47,12 +41,14 @@ class Marvin(BaseWebView):
 
     @route('/versions/')
     def get_versions(self):
-        vers = {'sess_vers': current_session['versions'], 'config_vers': config._mpldict.keys()}
+        vers = {'sess_vers': current_session['versions'], 'config_vers': config._allowed_releases.keys(),
+                'access': config.access, 'release': config.release, 'session_release': current_session['release']}
         return jsonify(result=vers)
 
     @route('/session/')
     def get_session(self):
-        return jsonify(result=dict(current_session))
+        extra = {'access': config.access}
+        return jsonify(result=dict(current_session, **extra))
 
     @route('/clear/')
     def clear_session(self):
@@ -93,8 +89,9 @@ class Marvin(BaseWebView):
                                             marvindb.datadb.Cube.plateifu).join(marvindb.datadb.PipelineInfo,
                                                                                 marvindb.datadb.PipelineVersion,
                                                                                 marvindb.datadb.IFUDesign).
-                     filter(marvindb.datadb.PipelineVersion.version == self._drpver).all())
+                     filter(marvindb.datadb.PipelineVersion.version == str(self._drpver)).all())
             out = [str(e) for l in cubes for e in l]
+
         out = list(set(out))
         out.sort()
         return jsonify(out)
@@ -104,40 +101,12 @@ class Marvin(BaseWebView):
         ''' Global selection of the MPL/DR versions '''
         args = av.manual_parse(self, request, use_params='index')
         version = args.get('release', None)
-        current_session['release'] = version
+        set_session_versions(version)
         drpver, dapver = config.lookUpVersions(release=version)
-        current_session['drpver'] = drpver
-        current_session['dapver'] = dapver
         out = {'status': 1, 'msg': 'Success', 'current_release': version,
                'current_drpver': drpver, 'current_dapver': dapver}
 
         return jsonify(result=out)
-
-    @route('/login/', methods=['GET', 'POST'], endpoint='login')
-    def login(self):
-        ''' login for trac user '''
-        try:
-            from inspection.marvin import Inspection
-        except ImportError as e:
-            from brain.core.inspection import Inspection
-
-        form = processRequest(request=request)
-        result = {}
-        username = form['username'].strip()
-        password = form['password'].strip()
-        auth = md5("{0}:AS3Trac:{1}".format(username, password).encode('utf-8')).hexdigest() if username and password else None
-        try:
-            inspection = Inspection(current_session, username=username, auth=auth)
-        except Exception as e:
-            result['status'] = 1
-            result['message'] = str(e)
-            current_session['loginready'] = True
-        else:
-            result = inspection.result()
-            current_session['loginready'] = inspection.ready
-            current_session['name'] = result.get('membername', None)
-
-        return jsonify(result=result)
 
     @route('/logout/', methods=['GET', 'POST'], endpoint='logout')
     def logout(self):
@@ -152,7 +121,46 @@ class Marvin(BaseWebView):
         if 'name' in current_session:
             name = current_session.pop('name')
 
+        request.environ['REMOTE_USER'] = None
+        config.access = 'public'
+        set_session_versions(config.release)
+        setGlobalSession()
+        logout_user()
+
         return redirect(url_for('index_page.Marvin:index'))
+
+    @route('/login/', methods=['GET', 'POST'], endpoint='login')
+    def login(self):
+        form = processRequest(request=request)
+        result = {}
+        username = form['username'].strip()
+        password = form['password'].strip()
+
+        # do nothing if already validated
+        if current_user.is_authenticated:
+            return jsonify(result=result)
+
+        # validate the user with htpassfile or trac username
+        is_valid, user, result = validate_user(username, password, request=request, session=current_session)
+
+        # get User only if valid
+        if is_valid:
+            user = get_db_user(username, password, dbsession=marvindb.session, user_model=marvindb.datadb.User, request=request)
+            if user and user.check_password(password):
+                login_user(user)
+
+        if is_valid:
+            result['status'] = 1
+            result['message'] = 'Login Successful!'
+            current_session['name'] = user.username
+            current_session['loginready'] = True
+            config.access = 'collab'
+            setGlobalSession()
+        else:
+            result['status'] = -1
+            result['message'] = 'Login {0} is not valid!'.format(username)
+
+        return jsonify(result=result)
 
 
 Marvin.register(index)