+++ CHANGELOG.rst
@@ -7,11 +7,30 @@ Marvin's Change Log
 Added
 ^^^^^
 - Added cheatsheet to docs.
+- New Web authentication using Flask-Login
+- New API authentication using Flask-JWT-Extended
+- Adds MPL-7 / DR15 datamodel
+- New config.access attribute indicating public or collab access
+- New config.login method to handle token-based login
+- New marvin.yml config file for customization of configuration options
+- Adds User table into the history schema of mangadb.  Tracks user logins.
+- ``Map`` has a new method ``getSpaxel`` to retrieve an ``Spaxel`` using the parent ``Maps``.
 
 Changed
 ^^^^^^^
+- Integrated datamodel plotting params into actual datamodel structures
+- Moved netrc checks into the Brain
+- Expanded sasurl into public and collab urls
+- Changes personal emails to sdss helpdesk email in web
+- Added rawsql and return_params columns to history.query table in mangadb
 - Extra keyword arguments passed to ``Spectrum.plot`` are now forwarded to ``Axes.plot``.
 - Tools (e.g., ``Cube``, ``Maps``) can now be accessed from the ``marvin`` namespace (e.g., ``marvin.tools.Cube`` or ``marvin.tools.cube.Cube``).
+
+Fixed
+^^^^^
+- Issue `#421 <https://github.com/sdss/marvin/issues/421>`_ - query returning spaxel props returns wrong total count
+- Bugfix - Python 3 xrange syntax bug in buildImageDict
+- ``Bin._create_spaxels`` instantiating spaxels with the wrong ``(i,j)`` values for the bin. The ``(i, j)`` values from the ``binid`` map were being swapped twice before sending them to ``SpaxelBase`` (:issue:`457`).
 
 Refactored
 ^^^^^^^^^^